move topology to mod
